Make ThreadKey and MessageKey not opaque
Opaque types hide their underlying structure when they are used outside
the lexical scope that they were defined in. This means that arbitrary
values that would otherwise be able to be coerced into that type, in
this case arbitrary strings, cannot be. An opaque type validates that a
module only be given a value that it itself gave the consumer: in this
case, it would only be possible to key into `messages` with a key
retrieved from a `Thread`.

This would be great in a virtual-DOM situation like React, where
message and thread keys were bound to event handlers in the JS codebase.
It breaks down when we consider binding keys to DOM elements as data
attributes, or the like, as those are always retrieved as strings.
Opaque types would cause more trouble than they’re worth, so I’m
removing them.